Lab2: timer and momentum conservation added

+++ Lab2/Particle.hpp
@@ -5,7 +5,7 @@
 class Particle
 {
 public:
-  float radius;
+  double radius;
   Vector pos;
   Vector vel;
 
@@ -13,22 +13,22 @@ public:
   //Particle(const Particle &val) = default;
   //Particle &operator=(const Particle & val) = default;
 
-  Particle(const float radius, const Vector &pos, const Vector &vel = Vector()) :
+  Particle(const double radius, const Vector &pos, const Vector &vel = Vector()):
     radius(radius), pos(pos), vel(vel) {}
 
   void move(const Vector &d)
   {
     pos = pos + d;
   }
-  void applyForce(const Vector &force, const float time)
+  void applyForce(const Vector &force, const double time)
   {
     vel = vel + force * time;
   }
-  void iter(const float time)
+  void iter(const double time)
   {
     pos = pos + vel * time;
   }
-  friend float dist(const Particle &a, const Particle &b)
+  friend double dist(const Particle &a, const Particle &b)
   {
     return dist(a.pos, b.pos);
   }
